requirements.txt
fast_plate_ocr==0.3.0
inference==0.49.5
inference[transformers]
inference[yolo-world]
inference[grounding-dino]
inference[gaze]
inference[clip]
inference[sam]
inference-sdk==0.49.5
numpy==2.2.5
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
supervision==0.25.1
python-dotenv
typer[all]==0.12.5
rich==13.0.1
